Animaatiot siirretty oikeisiin paikkoihin (Visualisointi ja IVisualisointi). OmaMoottorissa muokattu pieni logiikka fix animaatioon.
