+++ src/main/java/view/SimulaattorinGUI.java
@@ -37,22 +37,9 @@ public class SimulaattorinGUI extends Application implements ISimulaattorinUI {
 
     //Kontrollerin esittely (tarvitaan käyttöliittymässä)
     private IKontrolleriForV kontrolleri;
-    private int customerCount = 0;
 
     // Heitin root borderpanen tänne, käytän sitä useammassa metodissa
     private BorderPane root;
-
-    // Animaatio shittiä
-
-    // Lista varatuista sijainneista
-    private List<Point2D> occupiedPositions = new ArrayList<>();
-
-    // Asiakkaan koko, minimi välimatka kun luodaan asiakkaita, että ei mene päällekkäin jne.
-    private static final double CUSTOMER_RADIUS = 10;
-    private static final double MIN_DISTANCE = 3 * CUSTOMER_RADIUS;
-    private static final double areaWidth = 100;
-    private static final double areaHeight = 100;
-
 
     // Käyttöliittymäkomponentit:
     private TextField aika;
@@ -483,7 +470,7 @@ public class SimulaattorinGUI extends Application implements ISimulaattorinUI {
 
             labelBox.getChildren().add(grid);
 
-            naytto = new Visualisointi(309, 200);
+            naytto = new Visualisointi(309, 200, root);
 
             // Canvas skaalaus ei toimi vielä kunnolla !!
             Canvas canvas = (Canvas) naytto;
@@ -655,84 +642,6 @@ public class SimulaattorinGUI extends Application implements ISimulaattorinUI {
         DecimalFormat formatter = new DecimalFormat("#0.00");
         this.ETkokonaisAika.setText(formatter.format(totalTime));
     }
-
-    // Animaatio hommelit, nää varmaa vois laittaa Visualisointi luokkaan ja muutenkin paremmin tehdä
-    // mutta nyt tällänen jonkinlaine ratkaisu. Tässä on siis asiakkaiden piirtäminen ja liikuttaminen
-
-    @Override
-    public void drawCustomer(int id, double areaX, double areaY) {
-        double x = areaX;
-        double y = areaY;
-
-        // Check for overlaps and adjust position if necessary
-        boolean overlap;
-        do {
-            overlap = false;
-            for (Point2D position : occupiedPositions) {
-                if (position.distance(x, y) < MIN_DISTANCE) {
-                    overlap = true;
-                    x += MIN_DISTANCE; // Adjust x position
-                    if (x > areaX + areaWidth - CUSTOMER_RADIUS) { // If x exceeds area width, reset x and adjust y
-                        x = areaX;
-                        y += MIN_DISTANCE;
-                    }
-                    if (y > areaY + areaHeight - CUSTOMER_RADIUS) { // If y exceeds area height, reset y
-                        y = areaY;
-                    }
-                    break;
-                }
-            }
-        } while (overlap);
-
-        // Add the new position to the list of occupied positions
-        occupiedPositions.add(new Point2D(x, y));
-
-        Circle customer = new Circle(CUSTOMER_RADIUS, Color.BLUE);
-        customer.setId("customer-" + id);
-        customer.setCenterX(x);
-        customer.setCenterY(y);
-        root.getChildren().add(customer);
-    }
-
-    @Override
-    public void moveCustomer(int id, double toX, double toY, Runnable onFinished) {
-        Circle customer = (Circle) root.lookup("#customer-" + id);
-        if (customer != null) {
-            Point2D oldPosition = new Point2D(customer.getCenterX(), customer.getCenterY());
-            occupiedPositions.remove(oldPosition);
-
-            TranslateTransition transition = new TranslateTransition(Duration.seconds(1), customer);
-            transition.setToX(toX - customer.getCenterX());
-            transition.setToY(toY - customer.getCenterY());
-            transition.setOnFinished(event -> onFinished.run());
-            transition.play();
-        }
-    }
-
-    @Override
-    public void exitCustomer(int id, double toX, double toY) {
-        Circle customer = (Circle) root.lookup("#customer-" + id);
-        if (customer != null) {
-            Point2D oldPosition = new Point2D(customer.getCenterX(), customer.getCenterY());
-            occupiedPositions.remove(oldPosition);
-
-            TranslateTransition transition = new TranslateTransition(Duration.seconds(1), customer);
-            transition.setToX(toX - customer.getCenterX());
-            transition.setToY(toY - customer.getCenterY());
-
-            FadeTransition fadeTransition = new FadeTransition(Duration.millis(400), customer);
-            fadeTransition.setToValue(0);
-
-            SequentialTransition sequentialTransition = new SequentialTransition(transition, fadeTransition);
-            sequentialTransition.setOnFinished(event -> root.getChildren().remove(customer));
-            sequentialTransition.play();
-        }
-    }
-
-
-
-
-
 
     @Override
     public IVisualisointi getVisualisointi() {
